cast

+++ webrtc_patches/modules/video_coding/codecs/vp8/libvpx_vp8_decoder.cc
@@ -265,7 +265,7 @@ int LibvpxVp8Decoder::Decode(const EncodedImage& input_image,
   int ret;
 
   //Stegozoa
-  std::cout << "Address: " << rtpSession << std::endl;
+  std::cout << "Address: " << (uint64_t) rtpSession << std::endl;
   uint64_t data[2];
   data[0] = ssrc;
   data[1] = (uint64_t) rtpSession;
